ecom/requirements.txt
asgiref==3.7.2
Babel==2.14.0
Django==5.0.3
django-ckeditor==6.7.1
django-environ==0.11.2
django-filter==24.1
django-forms-dynamic==1.0.0
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
mysql==0.0.3
mysqlclient==2.2.4
phonenumbers==8.13.32
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
